% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect_header}
\alias{detect_header}
\title{Detect Comment Headers}
\usage{
detect_header(comment, n = 100, phrases = c("name", "date", "email"))
}
\arguments{
\item{comment}{character comment}

\item{n}{default = 100. checks the first n characters. Set to negative values or 0
to check the entire comment}

\item{phrases}{character vector of phrases to check for. Default is c('name', 'date', 'email')}
}
\value{
Logical vector of length comment
}
\description{
Attempts to identify if a comment has a header by checking for certain phrases
at the start of a comment. Phrases and first n characters of a comment are
converted to lower case.
}
\examples{
comms <- c('Name: Chris R is fun', 'Magrittr pipe is better than base pipe.',
'Date: 06/05/21 I love tibbles.')
detect_header(comms)
}
