% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icw.R
\name{icw}
\alias{icw}
\title{Inverse Covariance Weighting}
\usage{
icw(mat, reverse = NULL, standardize = TRUE)
}
\arguments{
\item{mat}{matrix}

\item{reverse}{column numbers to reverse. Default is NULL.}

\item{standardize}{Boolean to standardize columns. Default is TRUE.}
}
\value{
list with entries weight and index
}
\description{
Inverse Covariance Weighting
}
\examples{
set.seed(1)
mat <- matrix(rnorm(100), ncol = 5)
icw(mat)
}
