% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_colon}
\alias{replace_colon}
\title{Replace Colon in Start of Comments}
\usage{
replace_colon(comment, n = 100, replace = ".")
}
\arguments{
\item{comment}{character vector of comments}

\item{n}{default = 100. checks the first n characters. Set to negative values
to check the entire comment}

\item{replace}{character to replace ':' with. Default is '.'}
}
\value{
comment with colons replaced with \code{replace}
}
\description{
Replace Colon in Start of Comments
}
\examples{
set.seed(1)
comms <- lapply(1:10, function(x){
paste0(sample(c(LETTERS, ':'), size = 20, replace = TRUE), collapse = '')})
comms <- unlist(comms)
replace_colon(comms)
replace_colon(comms, n = 1)
replace_colon(comms, n = 0)
replace_colon(comms, n = 5)
}
