% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_mat}
\alias{standardize_mat}
\title{Standardize Matrix}
\usage{
standardize_mat(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix with same dimensions as mat
}
\description{
Standardize Matrix
}
\examples{
set.seed(1)
mat <- matrix(runif(16), 4, 4)
std_mat <- standardize_mat(mat)
}
